#182 (step 1): getting rid of "dynamic constants buffers" (use "RootDescriptor" instead): - removed dynamic constant buffers to simplify API and implementations (D3D12 mapped dynamic CBs into root descriptors under the hood, VK push descriptor is considered to be a modern and better approach to "bind a buffer with offset") - added dynamic "offset" to "RootDescriptor" - added "features.rootConstantsOffset" and "features.nonConstantBufferRootDescriptorOffset" which are not supported only in D3D11

+++ Source/D3D12/SharedD3D12.h
@@ -58,7 +58,7 @@ inline MemoryTypeInfo Unpack(const MemoryType& memoryType) {
 
 static_assert(sizeof(MemoryTypeInfo) == sizeof(MemoryType), "Must be equal");
 
-enum DescriptorHeapType : uint32_t {
+enum DescriptorHeapType : uint8_t {
     RESOURCE = 0,
     SAMPLER,
     MAX_NUM
